feat(toolbox): give announcements their own tab
The archive was a row of buttons tucked into settings, which is a poor
home for something that has to hold every announcement ever published.
It's now its own toolbox tab, listing each one in full as a dated card —
newest first, clearly separated, nothing to click before you can read it.

The announcement text moves out of the overlay into AnnouncementBody,
keyed by id, so the overlay and the archive render exactly the same thing
and a future announcement is a case here plus an archive entry.
